# Refactor normalize