--- a/R/normalize.R
+++ b/R/normalize.R
@@ -141,11 +141,12 @@ prep.step_normalize <- function(x, training, info = NULL, ...) {
 
 #' @export
 bake.step_normalize <- function(object, new_data, ...) {
-  res <- sweep(as.matrix(new_data[, names(object$means)]), 2, object$means, "-")
-  res[,names(object$sds)] <- sweep(res[,names(object$sds)], 2, object$sds, "/")
-  res <- tibble::as_tibble(res)
-  new_data[, names(object$means)] <- res
-  as_tibble(new_data)
+  for (column in names(object$means)) {
+    mean <- object$means[column]
+    sd <- object$sds[column]
+    new_data[[column]] <- (getElement(new_data, column) - mean) / sd
+  }
+  new_data
 }
 
 print.step_normalize <-