This PR aims to close https://github.com/tidymodels/recipes/issues/963.

The original goal was to make `step_normalize()` work when one variable is selected. I found that a small refactor of `bake.step_normalize()` not only fixed the issue but also provided a modest speed improvement. See below for benchmarking

# This PR

``` r
library(recipes)
library(nycflights13)

rec_spec <- recipe(~., data = flights) %>%
  step_zv(all_numeric_predictors()) %>%
  step_normalize(all_numeric_predictors())

bench::mark(
  rec_spec %>% prep() %>% bake(new_data = flights)
)
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 × 6
#>   expression                                            min   median `itr/sec`
#>   <bch:expr>                                       <bch:tm> <bch:tm>     <dbl>
#> 1 rec_spec %>% prep() %>% bake(new_data = flights)    291ms    302ms      3.31
#> # … with 2 more variables: mem_alloc <bch:byt>, `gc/sec` <dbl>

library(textrecipes)
library(friends)

friends_wide <- recipe(~ text, friends) %>%
  step_tokenize(text) %>%
  step_tokenfilter(text, max_tokens = 5000) %>%
  step_tfidf(text) %>%
  prep() %>%
  bake(new_data = NULL)

rec_spec <- recipe(~., data = friends_wide) %>%
  step_zv(all_numeric_predictors()) %>%
  step_normalize(all_numeric_predictors())

bench::mark(
  rec_spec %>% prep() %>% bake(new_data = friends_wide)
)
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 × 6
#>   expression                                                min median `itr/sec`
#>   <bch:expr>                                            <bch:t> <bch:>     <dbl>
#> 1 rec_spec %>% prep() %>% bake(new_data = friends_wide)   11.8s  11.8s    0.0850
#> # … with 2 more variables: mem_alloc <bch:byt>, `gc/sec` <dbl>
```

<sup>Created on 2022-05-06 by the [reprex package](https://reprex.tidyverse.org) (v2.0.1)</sup>

# Main

``` r
library(recipes)
library(nycflights13)

rec_spec <- recipe(~., data = flights) %>%
  step_zv(all_numeric_predictors()) %>%
  step_normalize(all_numeric_predictors())

bench::mark(
  rec_spec %>% prep() %>% bake(new_data = flights)
)
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 × 6
#>   expression                                            min   median `itr/sec`
#>   <bch:expr>                                       <bch:tm> <bch:tm>     <dbl>
#> 1 rec_spec %>% prep() %>% bake(new_data = flights)    544ms    544ms      1.84
#> # … with 2 more variables: mem_alloc <bch:byt>, `gc/sec` <dbl>

library(textrecipes)
library(friends)

friends_wide <- recipe(~ text, friends) %>%
  step_tokenize(text) %>%
  step_tokenfilter(text, max_tokens = 5000) %>%
  step_tfidf(text) %>%
  prep() %>%
  bake(new_data = NULL)

rec_spec <- recipe(~., data = friends_wide) %>%
  step_zv(all_numeric_predictors()) %>%
  step_normalize(all_numeric_predictors())

bench::mark(
  rec_spec %>% prep() %>% bake(new_data = friends_wide)
)
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 × 6
#>   expression                                                min median `itr/sec`
#>   <bch:expr>                                            <bch:t> <bch:>     <dbl>
#> 1 rec_spec %>% prep() %>% bake(new_data = friends_wide)   37.3s  37.3s    0.0268
#> # … with 2 more variables: mem_alloc <bch:byt>, `gc/sec` <dbl>
```

<sup>Created on 2022-05-06 by the [reprex package](https://reprex.tidyverse.org) (v2.0.1)</sup>